python/mcp/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "google-genai>=1.9.0",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
